+++ cabal-install/Distribution/Solver/Modular/Log.hs
@@ -4,7 +4,7 @@ module Distribution.Solver.Modular.Log
     ) where
 
 import Prelude ()
-import Distribution.Client.Compat.Prelude
+import Distribution.Solver.Compat.Prelude
 
 import Data.List as L
